refactor : EvalClient -> Client
